Cleanup command line interface (CLI) and remove stagnant /research directory
