--- research/All_indicators.csv
@@ -1,9 +1,0 @@
-,RSI,MFI,TSI,UO,Stoch,Stoch_Signal,WR,AO,MACD,MACD_signal,MACD_diff,EMA_fast,EMA_slow,Vortex_pos,Vortex_neg,Vortex_diff,Trix,Mass_index,CCI,DPO,KST,KST_sig,KST_diff,Ichimoku_a,Ichimoku_b,Aroon_up,Aroon_down,Aroon_ind,ATR,BBH,BBL,BBM,BBHI,BBLI,KCC,KCH,KCL,KCHI,KCLI,DCH,DCL,DCHI,DCLI,ADI,OBV,OBVM,CMF,FI,EM,VPT,NVI,DR,DLR,CR
-count,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0,54.0
-mean,0.2104820565010973,0.1509740435337152,0.14372582774203183,0.16273111067509932,0.18381329017684392,0.18081080150716244,0.18381329017684395,0.1810320876951954,0.1884083515449483,0.16274122686395992,0.11901177948053851,0.25495850957778987,0.25495850957778987,0.1587801326790828,-0.11127675655475372,0.04984858441118822,0.17252357017317757,0.007486681810315825,0.19603682392389518,-0.056103702339332026,0.15897067964874556,0.12417308112042205,0.12017413579503937,0.25245554307556384,0.2523461782341115,0.16351329411615634,-0.12508524827194473,0.16630871989246548,0.18342291403091407,0.2542797228114263,0.25430461623452694,0.25431753069184365,0.10973766387992209,-0.05780250636566367,0.25497845252407325,0.2549434619916608,0.25501144661371344,0.010620216408781694,0.01125091999165259,0.25442979204104504,0.254461348144139,0.14372148924844785,-0.08495006873735457,0.1188438480088641,0.10862929452736952,0.16010407604114238,0.13235095130726215,0.07263917976808297,0.034312490730126086,0.09478799179658669,-0.03570125387262867,0.1025573373097345,0.10532402513212934,0.2559708530221941
-std,0.3516631600953852,0.2546367612929211,0.26759388957461744,0.27168743260659556,0.3344788627757645,0.3358306102200431,0.3344788627757646,0.30579494086687714,0.3177188173508206,0.30635460008727594,0.2590482466844916,0.43664418106104697,0.43664418106104697,0.22530048644991305,0.31655342647479096,0.15490305517062994,0.32378246350502593,0.1443364357519492,0.34962714655472255,0.1907119188495354,0.3106174130694961,0.2912314211932694,0.2542509665549134,0.43813467667183226,0.43818525241070405,0.26715493556215536,0.29031529453256794,0.3153299325064766,0.35162322921365197,0.43720440061378824,0.4368571078542329,0.4370381527480025,0.1847382037696725,0.2256638358320669,0.43662940939345724,0.4367180527921492,0.4365366924182082,0.13856204039085102,0.13848398562183845,0.4370897481819405,0.4367898097913218,0.2284141034374641,0.2572651548608125,0.20253391252355957,0.2063326710228711,0.24186156262684952,0.21940731969851476,0.16686602048441468,0.13896584002389348,0.1889917982357096,0.16822551307448305,0.22968453386489968,0.2351065111108535,0.4360154007435417
-min,-0.6127112118441945,-0.4794084238677665,-0.42990356303069793,-0.39900172418878993,-0.5166234877075209,-0.6051657831094899,-0.5166234877075215,-0.4202102482691597,-0.40790531907301986,-0.36651383056913317,-0.5457253542435268,-0.18020616474964266,-0.18020616474964266,-0.38919898835128613,-0.625615546400303,-0.01651637726075065,-0.45590224161337817,-0.07506830668284292,-0.625615546400303,-0.5457253542435268,-0.3884379366347953,-0.29615687301792615,-0.4647205968753569,-0.1800472340253908,-0.17991011099216755,-0.5082502810056753,-0.8665521050684573,-0.8665521050684573,-0.3319746081459019,-0.18191871625975178,-0.17841410113539927,-0.18018326528747391,-0.2411499228793651,-0.4363069908225636,-0.18017905436249826,-0.18097005604281724,-0.17938177292067228,-0.05014049937165909,-0.05247555848328961,-0.18155271747894766,-0.17864786613159078,-0.33346759445533825,-0.5366382829224209,-0.2225024963429532,-0.27019999779772114,-0.3799192362803956,-0.3240566446887736,-0.11687147874774041,-0.05190106398227944,-0.17348278166263947,-0.3319746081459019,-0.2880330614249734,-0.32353236858653023,-0.180201334443733
-25%,-0.0016008620970026722,-0.0040817160092746,0.001473188020314264,0.013652712523384193,-0.012196480868384325,-0.013332854625039765,-0.012196480868384388,0.0008408120239150491,0.002387584502962517,0.004428867094256699,-0.004607452531816083,-0.0017557159546677191,-0.0017557159546677191,0.07531493742969948,-0.37857666061510864,-0.0017492920255717012,0.0005164732112094531,-0.048475059033419726,-0.006691344707539266,-0.12566614888647615,0.005011396006557426,0.0006543528772688764,-0.004950732613598293,-0.007908926807209987,-0.008371637753653155,0.008676119842951305,-0.35545073493648116,-0.002138401290028173,-0.030174665497036707,-0.0025801934433449815,-0.003342571230896426,-0.002359948390319904,0.003985004119665528,-0.18838736950681537,-0.0025868261209268193,-0.0025439638604102194,-0.002629921848388708,-0.028502295118534318,-0.027269769214786056,-0.002393912805419351,-0.003015609245308087,-0.007317037230941473,-0.22358247368226447,0.0027309837246771397,0.00037985341210564213,0.0028460374904703853,-0.0006481966486686387,-0.017390730142658885,-0.014669778212275107,0.001599066785938956,-0.1521578607029762,-0.0009131161848182596,-0.0016892221642079137,-0.0012362221009961612
-50%,0.1941271545231218,0.10841832095959561,0.020213050305456542,0.11556517942835598,0.11041214765591005,0.10725604150915591,0.11041214765591009,0.07104399660982957,0.09099373020326516,0.042770367797089474,0.0282235808108642,0.005007257003071492,0.005007257003071492,0.08693742220053789,-0.0950042537551917,0.010202066914639684,0.03143540970795909,-0.014590099710441957,0.14819975955038586,-0.03292037564847153,0.05226693958264668,0.016361365436738128,0.041397055806954554,-0.00031268781280878303,-0.0005136471394009458,0.057622027751171,-0.06261104534758938,0.06788569324823357,-0.0026676869585122187,0.0033367998505146844,0.0038654383925105004,0.0033901039302213073,0.06482990657391682,-0.05542859510699827,0.005448746922351021,0.00505911509638304,0.005681378403357399,-0.0008203616571591146,-0.0006602131464056688,0.00365537823178831,0.004123334606276022,0.11174268170013624,-0.08416181455547872,0.07905136733007675,0.05012831004859238,0.15298792460527405,0.07101335310907869,0.05853919041780767,0.008566357036909045,0.04942658831730429,-0.011188479533286465,0.017741057693994894,0.00903132239660832,0.006348594542798205
-75%,0.41986348800772993,0.3091626811571664,0.2765615555789307,0.29969595983261216,0.37348121242091503,0.3787008285482535,0.3734812124209151,0.3818404745337085,0.3494516430477292,0.23427706874608067,0.2309505579211004,0.5927505362987808,0.5927505362987808,0.2843274417133523,0.14316343093471776,0.032686222754550194,0.277976632761088,0.02685532250123188,0.39796331034246585,-0.0014335222659658015,0.24499703362077482,0.13263103234091772,0.30394917247870556,0.594575223911342,0.594873416090042,0.33794746943483234,0.012922531157267962,0.39555111618971633,0.594540454782695,0.596763053833175,0.5896917799231365,0.5932721228488271,0.18919820362043455,0.006580416343205895,0.5926726376390298,0.5941611484549683,0.5911701160621715,0.0012009019461966053,0.0013015185747129224,0.5960236087629351,0.59007121736674,0.23223202599057488,0.006476196259995918,0.21864144742411784,0.17983453363406302,0.328987858162071,0.2888887978241029,0.08759795363045222,0.045545732299225726,0.12154354711166895,-0.0011264196573303393,0.13908928702645632,0.14290802801202948,0.5921094788172189
-max,1.0,1.0,1.0,1.0,1.0000000000000002,1.0,1.0000000000000002,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0
